update to 1.6.3

+++ src/services/Git.php
@@ -24,6 +24,12 @@ class Git{
 		$clearworkspace = isset($parameters['clearworkspace'])?$parameters['clearworkspace']:false;
 		$directory = isset($parameters['directory'])?$parameters['directory']:'';
 
+		$projectid = isset($parameters['projectid'])?$parameters['projectid']:null;
+		$token = isset($parameters['token'])?$parameters['token']:null;
+
+		$urlParse = parse_url($url);
+
+		
 		if( $clearworkspace === true ){
 			File::rrmdir( $workspace );
 		}
@@ -40,7 +46,31 @@ class Git{
 		$repository = end($urlArray);
 		$repositoryBranch = $repository.'-'.$branch;		
 
-		$url = $url. '/archive/'.$branch.'.zip';
+		$modeGet = null;
+
+		if($urlParse['host'] == 'github.com'){
+			$url = $url. '/archive/'.$branch.'.zip';
+			$modeGet = 'github';
+		}
+		if($urlParse['host'] == 'gitlab.com'){
+			$modeGet = 'gitlab';
+
+			if($projectid === null){
+				die('OnService-Git: missing project ID');
+			}
+			if($token === null){
+				die('OnService-Git: missing token');
+			}
+
+			$url = 'https://gitlab.com/api/v4/projects/'.$projectid.'/repository/archive?private_token='.$token;
+
+			// https://gitlab.com/api/v3/projects/11248159/repository/archive?private_token=n1EQF9NYzCHSmvQqxvh9
+
+			// $url = $url. '/-/archive/'.$branch.'/'.$repositoryBranch.'.zip';
+		}
+
+
+		
 
 		$data = array(
 			'url'=>$url,					
@@ -50,17 +80,79 @@ class Git{
 		if($username != null && $password != null)
 		$data['autenticate'] = "$username:$password";
 
+
 		$response = HttpCon::request($data);
+
+
 		if($response == false) return false;
 
-		$filename = $branch.'.zip';
-		$filepath = $workspace.DIRECTORY_SEPARATOR.$filename;
-		file_put_contents($filepath, $response);
-		$filetarget = $workspace;
-		Zip::extract($filepath,$filetarget);
-		File::rcopy($filetarget.DIRECTORY_SEPARATOR.$repositoryBranch.DIRECTORY_SEPARATOR.$directory, $filetarget.'/');
-		unlink($filepath);
-		File::rrmdir( $filetarget.DIRECTORY_SEPARATOR.$repositoryBranch.DIRECTORY_SEPARATOR );
+		if($modeGet == 'github'){
+			$filename = $branch.'.zip';
+			$filepath = $workspace.DIRECTORY_SEPARATOR.$filename;
+
+			file_put_contents($filepath, $response);		
+			$filetarget = $workspace;
+
+			Zip::extract($filepath,$filetarget);
+			File::rcopy($filetarget.DIRECTORY_SEPARATOR.$repositoryBranch.DIRECTORY_SEPARATOR.$directory, $filetarget.'/');
+			unlink($filepath);
+			File::rrmdir( $filetarget.DIRECTORY_SEPARATOR.$repositoryBranch.DIRECTORY_SEPARATOR );
+		}
+
+		if($modeGet == 'gitlab'){
+			$filename = $branch;
+			$filepath = $workspace.DIRECTORY_SEPARATOR.$filename;
+
+			//delete temp.tar.gz
+			if (file_exists($filepath.'.tar.gz')) {
+			    unlink($filepath.'.tar.gz');
+			}
+
+			//delete temp.tar
+			if (file_exists($filepath.'.tar')) {
+			    unlink($filepath.'.tar');
+			}
+
+			file_put_contents($filepath.'.tar.gz', $response);		
+			$filetarget = $workspace;
+			
+			$p = new \PharData($filepath.'.tar.gz');
+			$p->decompress(); // creates files.tar
+
+		
+			$phar = new \PharData($filepath.'.tar');
+			$phar->extractTo($workspace.DIRECTORY_SEPARATOR, null, true); // extract all files, and overwrite
+			
+			//delete temp.tar.gz
+			if (file_exists($filepath.'.tar.gz')) {
+			    unlink($filepath.'.tar.gz');
+			}
+
+			//delete temp.tar
+			if (file_exists($filepath.'.tar')) {
+			    unlink($filepath.'.tar');
+			}
+
+			$dirArray = scandir($workspace.DIRECTORY_SEPARATOR);
+
+			$file = null;
+			foreach ($dirArray as $key => $value) {
+				if(strpos($value, $repositoryBranch)!== false){
+					$file = $value;
+					break;
+				}
+			}
+
+			File::rcopy($workspace.DIRECTORY_SEPARATOR.$file.DIRECTORY_SEPARATOR.$directory, $workspace.DIRECTORY_SEPARATOR);
+			File::rrmdir( $workspace.DIRECTORY_SEPARATOR.$file );
+
+		}
+
+
+	
+		
+		
+		
 
 		return true;
